+++ docs/en/CHANGELOG.md
@@ -37,6 +37,7 @@
 - change post acceleration to true false
 - rework clicklock
 - change refill and harvest time, also for the bottom mode
+- add pause/restart
 
 ## 0.5.3 (2025-06-21)